Add solution for problem `G - MinOr Tree` Link: https://codeforces.com/contest/1624/problem/G Tags: bitmasks, dfs and similar, dsu, graphs, greedy, *1900 Ref: https://codeforces.com/contest/1624/submission/142320684

+++ codeforces/1624/G.cpp
@@ -6,7 +6,7 @@
 using namespace std;
 #define dbg(...)
 #endif
-using ll = int64_t;
+using ll = int;
 constexpr auto sz = [](const auto &container) -> ll { return ll(container.size()); };
 #define all(x) (x).begin(), (x).end()
 template <class T, class U = T>
